Alter game_exists check.
The process will now check for a BoardGameGeek ID in the index, and then attempt to retrieve a WordPress post,
if one exists, with a fallback to the pre-existing query. If no post ID is found, the BoardGameGeek ID is unset
from the index to facilitate with keeping it up to date.