Implement logic to check for valid usernames on BoardGameGeek.
The plugin will now save a boolean value for verified BGG usernames. On load or update of the settings page, if the verified
value is false, the plugin will save the new verified value. If the value is set to false, it will attempt to query the BGG
API for the username. Users which return an ID in the response are considered valid and will not render a notification.